修复批量删除bug 当没有勾选id时出现空指针错误

+++ src/com/his1/controller/PatientController.java
@@ -48,7 +48,9 @@ public class PatientController {
 	
 	@RequestMapping("/delPatient")
 	public String delPatient(Integer[] patientId) throws Exception{
-		
+		if(patientId==null){
+			return "redirect:queryPatient.action";
+		}
 		for (Integer integer : patientId) {
 			patientService.deletePatient(integer);
 		}
